+++ tests/test_aiosonic.py
@@ -62,70 +62,6 @@ async def test_simple_get_aiodns(http_serv, mocker):
         assert await res.text() == "Hello, world"
 
 
-@pytest.mark.asyncio
-@pytest.mark.timeout(15)
-async def test_get_python(http2_serv):
-    """Test simple get."""
-    url = http2_serv
-
-    connector = TCPConnector(timeouts=Timeouts(sock_connect=3, sock_read=4))
-    async with aiosonic.HTTPClient(connector) as client:
-        res = await client.get(
-            url,
-            verify=False,
-            headers={
-                "user-agent": (
-                    "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.14; rv:70.0)"
-                    " Gecko/20100101 Firefox/70.0"
-                )
-            },
-            http2=True,
-        )
-        assert "Hello World" in await res.text()
-
-
-@pytest.mark.asyncio
-@pytest.mark.timeout(15)
-async def test_post_http2(http2_serv):
-    """Test simple post."""
-    url = f"{http2_serv}/post"
-
-    # connector = TCPConnector(timeouts=Timeouts(sock_connect=3, sock_read=4))
-    connector = TCPConnector()
-    async with aiosonic.HTTPClient(connector) as client:
-        res = await client.post(
-            url,
-            json={"foo": "bar"},
-            verify=False,
-            http2=True,
-        )
-        assert "Hello World" in await res.text()
-
-
-@pytest.mark.asyncio
-@pytest.mark.timeout(15)
-async def test_get_http2(http2_serv):
-    """Test simple get to node http2 server."""
-    url = http2_serv
-    connector = TCPConnector(timeouts=Timeouts(sock_connect=3, sock_read=4))
-
-    async with aiosonic.HTTPClient(connector) as client:
-        res = await client.get(url, verify=False)
-        assert res.status_code == 200
-        assert "Hello World" == await res.text()
-
-
-@pytest.mark.asyncio
-@pytest.mark.timeout(15)
-async def test_method_lower(http2_serv):
-    """Test simple get to node http2 server."""
-    url = http2_serv
-    async with aiosonic.HTTPClient() as client:
-        res = await client.request(url, method="get", verify=False)
-        assert res.status_code == 200
-        assert "Hello World" == await res.text()
-
-
 class MyConnection(Connection):
     """Connection to count keeped alives connections."""
 
@@ -385,43 +321,6 @@ async def test_pool_acquire_timeout(http_serv, mocker):
                 client.get(url),
                 client.get(url),
             )
-
-
-@pytest.mark.asyncio
-async def test_simple_get_ssl(http2_serv):
-    """Test simple get with https."""
-    url = http2_serv
-
-    async with aiosonic.HTTPClient() as client:
-        res = await client.get(url, verify=False)
-        assert res.status_code == 200
-        assert await res.text() == "Hello World"
-
-
-@pytest.mark.asyncio
-async def test_simple_get_ssl_ctx(http2_serv):
-    """Test simple get with https and ctx."""
-    url = http2_serv
-
-    ssl_context = ssl.create_default_context(
-        ssl.Purpose.SERVER_AUTH,
-    )
-    ssl_context.check_hostname = False
-    ssl_context.verify_mode = ssl.CERT_NONE
-    async with aiosonic.HTTPClient() as client:
-        res = await client.get(url, ssl=ssl_context)
-        assert res.status_code == 200
-        assert await res.text() == "Hello World"
-
-
-@pytest.mark.asyncio
-@pytest.mark.timeout(2)
-async def test_simple_get_ssl_no_valid(http2_serv):
-    """Test simple get with https no valid."""
-    url = http2_serv
-    async with aiosonic.HTTPClient() as client:
-        with pytest.raises(ssl.SSLError):
-            await client.get(url)
 
 
 @pytest.mark.asyncio
@@ -717,37 +616,6 @@ async def test_wait_connections_busy_timeout(mocker):
     connector = TCPConnector(pool_cls=CyclicQueuePool)
     async with aiosonic.HTTPClient(connector) as client:
         assert not await client.wait_requests(0)
-
-
-@pytest.mark.asyncio
-@pytest.mark.timeout(15)
-async def test_get_image(http2_serv):
-    """Test get image."""
-    url = f"{http2_serv}/sample.png"
-
-    async with aiosonic.HTTPClient() as client:
-        res = await client.get(url, verify=False)
-        assert res.status_code == 200
-        assert res.chunked
-        with open("tests/sample.png", "rb") as _file:
-            assert (await res.content()) == _file.read()
-
-
-@pytest.mark.asyncio
-@pytest.mark.timeout(15)
-async def test_get_image_chunked(http2_serv):
-    """Test get image chunked."""
-    url = f"{http2_serv}/sample.png"
-
-    async with aiosonic.HTTPClient() as client:
-        res = await client.get(url, verify=False)
-        assert res.status_code == 200
-        assert res.chunked
-        filebytes = b""
-        async for chunk in res.read_chunks():
-            filebytes += chunk
-        with open("tests/sample.png", "rb") as _file:
-            assert filebytes == _file.read()
 
 
 @pytest.mark.asyncio